backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
asyncpg
python-multipart
gunicorn
email-validator
requests
httpx
Pillow
aiofiles
# bcrypt
duckduckgo-search
pyjwt
langchain-openai
langgraph
langchain-core
redis
supabase
